Refactor cli args
